backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
aiohttp
pydantic-settings
jinja2
groq